Added linter config to dots as well as an initial mackup config
Added rules to eslintrc to remove errors when using Jasmine, mocha (need
to add chai) to the env vars for my config also added it to dots so it
is version controlled.
Also added an initial rudimentary mackup config file to dots and finally
edited vimrc to remove ale which was not working to add Neomake, which
is working.