refactor(lspconfig): add convert_config helper
this is an exercise in over complication and over engineering/
ultimately was just fun for me. I created a "switch" statement of sorts
to handle the various types that are specified in the servers table.

A simpler solution would just be to be more consistent about the structure
as I was doing previously but this was fun 🤷