refactor(plugins): lazyload lspconfig
for reasons I don't understand it takes about 20ms to load not sure if
this is nvim-lsp-installer or some sort of issue in my local setup etc.
no idea but I've had to defer it's load time which seems to save me
around 40ms since it also interacts with nvim-treesitter causing that to
load very slowly as well.