refactor(packer)!: improve use local for contributions
Allow using "use_local" with "local_enabled" as a way of preferring
a local dev plugin to the upstream version as long as it is available on
the system. This is easier to use than an environment variable that
enables all dev plugins at once which is super noisy. It's much better
whilst contributing to have one enabled then remove use local otherwise
or just disable it