refactor(lsp): add diagnostic line highlighting
This re-adds a feature I've played with on and off for a while i.e.
highlighting a line with a diagnostic as the colour of the diagnostic.
I find this is a more subtle/less noisy indicator than virtual text but
can frankly be used with virtual text.

It's often the case that virtual text is cut off and it doesn't really
provide enough information anyway, so just knowing where the errors are
more visibly is often enough