refactor(lsp): track clients and features per buffer
Whilst using autocommands with lsp attach and detach there are quite
a few gotchas essentially around when the methods run if there is valid
client or the client has detached. My current solution for this is to
create a buffer-local table that tracks features, their augroup ids as
well as the clients that have that capability.

this is so when a client detaches I can remove augroups for capabilities
that are no longer supported.

I feel like this should be easier or native to upstream not implemented
like this, so maybe something that can be upstream but it's hard to say,
seems like they won't accept something like this