fix(statusline): don't use vim.fn.expand in dir_env
For completely unclear reasons when using vim.fn.expand in the dir_env
function as part of trying to normalize a path to compare it. The
commandline no longer completes using tab but just cycles two items over
and over. I've found that switching to the vim.fs.method fixes this