refactor(lsp)!: simplify lsp autocommand creation
I was previously going some fairly complex checks to make sure that if
an lsp server detached I wouldn't be running the autocommands for that
server any more if no server remained that supported that functionality.
This is involved some fairly byzantine logic for a situation that is
frankly, a bug. LSP should connect to a buffer and remain connect unless
something goes wrong, so I shouldn't need complex logic to prevent that.